identity field types were changed to bigint

+++ publishable/database/migrations/2017_11_05_060225_create_admin_settings_table.php
@@ -14,10 +14,10 @@ class CreateAdminSettingsTable extends Migration
     public function up()
     {
         Schema::create('admin_settings', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name', 191);
 	        $table->string('key', 191);
-	        $table->integer('admin_setting_group_id')
+	        $table->bigInteger('admin_setting_group_id')
 	              ->unsigned()
 	              ->nullable();
