+++ publishable/database/migrations/2018_01_15_085827_create_user_role_table.php
@@ -14,23 +14,23 @@ class CreateUserRoleTable extends Migration
     public function up()
     {
         Schema::create('user_role', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
 
             $table->foreign('user_id', 'user_role_user_id_foreign')
-                  ->references('id')
-                  ->on('users')
-                  ->onDelete('cascade');
+                ->references('id')
+                ->on('users')
+                ->onDelete('cascade');
 
-	        $table->integer('role_id')->unsigned();
+            $table->bigInteger('role_id')->unsigned();
 
-	        $table->foreign('role_id', 'user_role_role_id_foreign')
-	              ->references('id')
-	              ->on('roles')
-	              ->onDelete('cascade');
+            $table->foreign('role_id', 'user_role_role_id_foreign')
+                ->references('id')
+                ->on('roles')
+                ->onDelete('cascade');
 
-	        $table->json('meta')->nullable();
+            $table->json('meta')->nullable();
 
-	        $table->timestamps();
+            $table->timestamps();
 
         });
     }
@@ -42,10 +42,10 @@ class CreateUserRoleTable extends Migration
      */
     public function down()
     {
-	    Schema::table('user_role', function (Blueprint $table) {
-	    	$table->dropForeign('user_role_user_id_foreign');
-		    $table->dropForeign('user_role_role_id_foreign');
-	    });
+        Schema::table('user_role', function (Blueprint $table) {
+            $table->dropForeign('user_role_user_id_foreign');
+            $table->dropForeign('user_role_role_id_foreign');
+        });
 
         Schema::dropIfExists('user_role');
     }
